Hack to make tags balance correctly
This extension doesn't parse HTML correctly.  If you have `<foo><foo>bar</foo>baz</foo>`, it will perceive the interior of the outermost tag to stop at the FIRST `</foo>`, and thus screw up the recursion entirely.  So I extended it to accept not just the tag "htmltag" but any tag of the form "htmltag0" through "htmltag50" and will encode tags with increasing numbers with the depth of nesting.  This is a hack, but the MediaWiki bug tracker has (literally) over a hundred thousand entries, scores on the parser alone, so I despair of getting any help there.